+++ src/components/form/index.ts
@@ -7,8 +7,7 @@ export * from "./Select/ControlledStringSelect";
 export * from "./Select/ControlledSelect";
 export * from "./Select/StringSelect";
 export * from "./Select/Select";
-export * from "./TextInput/InputText";
-export * from "./TextInput/StateInputText";
 export * from "./TextInput/TextInput";
+export * from "./TextInput/StateInputText";
 export * from "./TimeInput/TimeInput";
 export * from "./TimeInput/TimeInputV2";
